+++ SpeckleUnrealProject/Plugins/SpeckleUnreal/Source/SpeckleUnreal/Public/SpeckleUnreal.h
@@ -3,44 +3,13 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Modules/ModuleManager.h"
 #include "LevelEditor.h"
-#include "SlateBasics.h"
-#include "Interfaces/IMainFrameModule.h"
-#include "FSpeckleEditorCommands.h"
 
 
 class FSpeckleUnrealModule : public IModuleInterface
 {
 public:
 
-	/** IModuleInterface implementation */
 	virtual void StartupModule() override;
 	virtual void ShutdownModule() override;
-
-	TSharedPtr<FExtender> ToolbarExtender;
-	TSharedPtr<const FExtensionBase> Extension;
-	TSharedPtr<SWindow> SpeckleEditorWindow;
-	FOnWindowClosed OnSpeckleWindowClosed;
-
-	
-	void SpeckleButtonListener();
-
-	UFUNCTION()
-	void OnEditorWindowClosed(const TSharedRef<SWindow>&);
-	
-	void AddToolbarExtension(FToolBarBuilder &builder)
-	{
-		FSlateIcon IconBrush = FSlateIcon(FEditorStyle::GetStyleSetName(), "LevelEditor.ViewOptions", "LevelEditor.ViewOptions.Small");
-		
-		builder.AddToolBarButton
-		(
-			FSpeckleEditorCommands::Get().SpeckleToolBarBtn,
-			NAME_None,
-		    FText::FromString("Speckle Editor"), 
-		    FText::FromString("Click to open Speckle Window"),
-		    IconBrush,
-		    NAME_None
-        );
-	};
 };
