+++ SpeckleUnrealProject/Plugins/SpeckleUnreal/SpeckleUnreal.uplugin
@@ -17,7 +17,7 @@
 	"Modules": [
 		{
 			"Name": "SpeckleUnreal",
-			"Type": "Editor",
+			"Type": "Runtime",
 			"LoadingPhase": "Default"
 		},
 		{
